requirements.txt
torch
python-chess
pygame
numpy 